// adjuster-portal-ui/src/app/models/files-user.model.ts
export class UserPersonalFile {
  id?: string;
  user_id?: string;
  name: string;
  description?: string;
  state?: string;
  expiration_date?: string;
  type?: string;
  size?: number;
  path?: string;
  can_be_removed?: boolean = true;

  created_by: {
    first_name: string;
    last_name: string;
  }
  updated_by: {
    first_name: string;
    last_name: string;
  }
  created_at: Date;
  updated_at: Date;

}